Refactor navigation stacks
