+++ RevenueCatUI/CustomerCenter/Views/CompatibilityNavigationStack.swift
@@ -1,0 +1,47 @@
+//
+//  Copyright RevenueCat Inc. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
+//  CompatibilityNavigationStack.swift
+//
+//
+//  Created by Cody Kerns on 8/15/24.
+//
+
+
+#if CUSTOMER_CENTER_ENABLED
+
+#if os(iOS)
+
+import SwiftUI
+
+@available(iOS 15.0, *)
+@available(macOS, unavailable)
+@available(tvOS, unavailable)
+@available(watchOS, unavailable)
+struct CompatibilityNavigationStack<Content:View>: View {
+    
+    @ViewBuilder var content: Content
+
+    var body: some View {
+        if #available(iOS 16.0, *) {
+            NavigationStack {
+                content
+            }
+        } else {
+            NavigationView {
+                content
+            }
+        }
+    }
+    
+}
+
+#endif
+
+#endif
